% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{odata_function}
\alias{odata_function}
\title{Make an OData function available to R}
\usage{
odata_function(url, metadata = c("none", "minimal", "all"), ...)
}
\arguments{
\item{url}{Which url to fetch data from}

\item{metadata}{Whether and how metadata is included}

\item{...}{
  Arguments passed on to \code{\link[=retrieve_data]{retrieve_data}}
  \describe{
    \item{\code{httr_args}}{List of additional arguments passed on to httr::GET}
    \item{\code{jsonlite_args}}{List of additional arguments passed on to jsonlite::fromJSON}
  }}
}
\value{
An R function
}
\description{
This turns an OData function into an R function
Parameters are serialized to json.
Scalar arguments should be passed as atomic vectors.
Array or object arguments should be passed as list.
}
\seealso{
Other retrieve: 
\code{\link{retrieve_all}()},
\code{\link{retrieve_data}()},
\code{\link{retrieve_one}()}
}
\concept{retrieve}
