% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve data}
\usage{
retrieve_data(
  url,
  metadata = c("none", "minimal", "all"),
  httr_args = list(),
  jsonlite_args = list()
)
}
\arguments{
\item{url}{Which url to fetch data from}

\item{metadata}{Whether and how metadata is included}

\item{httr_args}{List of additional arguments passed on to httr::GET}

\item{jsonlite_args}{List of additional arguments passed on to jsonlite::fromJSON}
}
\value{
Data including metadata
}
\description{
Retrieve data
}
\examples{
\dontrun{
url <- "https://services.odata.org/V4/TripPinServiceRW"
retrieve_data(url)
}
}
\seealso{
Other retrieve: 
\code{\link{odata_function}()},
\code{\link{retrieve_all}()},
\code{\link{retrieve_one}()}
}
\concept{retrieve}
